Fixed some warnings

+++ src/Engine/TMX/Tileset.cpp
@@ -51,8 +51,8 @@ unsigned int Tileset::GetTileHeight() const
 
 const umath::rectangle Tileset::GetTile(unsigned int localId) const
 {
-	const int wTiles = m_texture->GetSize().x / m_tileWidth;
-	const int hTiles = m_texture->GetSize().y / m_tileHeight;
+	const unsigned int wTiles = static_cast<unsigned int>(m_texture->GetSize().x) / m_tileWidth;
+    const unsigned int hTiles = static_cast<unsigned int>(m_texture->GetSize().y) / m_tileHeight;
 
 	int x = localId % wTiles * m_tileWidth;
 	int y = localId / wTiles * m_tileHeight;
